Fixed cast warnings: explicitly cast strlen result to int

+++ src/FMDatabase.m
@@ -160,7 +160,7 @@
         return NO;
     }
     
-    int rc = sqlite3_rekey(db, [key UTF8String], strlen([key UTF8String]));
+    int rc = sqlite3_rekey(db, [key UTF8String], (int)strlen([key UTF8String]));
     
     if (rc != SQLITE_OK) {
         NSLog(@"error on rekey: %d", rc);
@@ -179,7 +179,7 @@
         return NO;
     }
     
-    int rc = sqlite3_key(db, [key UTF8String], strlen([key UTF8String]));
+    int rc = sqlite3_key(db, [key UTF8String], (int)strlen([key UTF8String]));
     
     return (rc == SQLITE_OK);
 #else
